All MicroServices Added and working

+++ backend/rest-service/src/main/java/com/backend/restservice/repository/AlquilerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.backend.restservice.entity.Alquiler;
 
 @Repository
-public interface AlquilerRepository extends JpaRepository<Alquiler, Long> {
+public interface AlquilerRepository extends JpaRepository<Alquiler, Integer> {
 
 }
